fix: address review comments - extract shared quota detection utility and fix translation typos
