+++ src/services/code-index/embedders/openai.ts
@@ -8,6 +8,7 @@ import {
 	MAX_BATCH_RETRIES as MAX_RETRIES,
 	INITIAL_RETRY_DELAY_MS as INITIAL_DELAY_MS,
 } from "../constants"
+import { isInsufficientQuotaError } from "./utils/quota-detection"
 import { getModelQueryPrefix } from "../../../shared/embeddingModels"
 import { t } from "../../../i18n"
 
@@ -142,7 +143,7 @@ export class OpenAiEmbedder extends OpenAiNativeHandler implements IEmbedder {
 				const hasMoreAttempts = attempts < MAX_RETRIES - 1
 
 				// Add quota detection
-				const isQuotaError = this.isInsufficientQuotaError(error)
+				const isQuotaError = isInsufficientQuotaError(error)
 
 				if (isRateLimitError && !isQuotaError && hasMoreAttempts) {
 					const delayMs = INITIAL_DELAY_MS * Math.pow(2, attempts)
@@ -198,32 +199,5 @@ export class OpenAiEmbedder extends OpenAiNativeHandler implements IEmbedder {
 		return {
 			name: "openai",
 		}
-	}
-
-	/**
-	 * Detects if an error is due to insufficient quota/credits
-	 * @param error The error object to check
-	 * @returns True if the error indicates insufficient quota
-	 */
-	private isInsufficientQuotaError(error: any): boolean {
-		if (error?.status !== 429) return false
-
-		const errorMessage =
-			error?.message?.toLowerCase() ||
-			error?.response?.data?.error?.message?.toLowerCase() ||
-			error?.error?.message?.toLowerCase() ||
-			""
-
-		const quotaKeywords = [
-			"insufficient_quota",
-			"insufficient quota",
-			"quota exceeded",
-			"insufficient funds",
-			"billing",
-			"payment required",
-			"credits",
-		]
-
-		return quotaKeywords.some((keyword) => errorMessage.includes(keyword))
 	}
 }
